OWA: Move the attachment requests out of the calendar
They are not calendar-specific, and the contacts need them, too,
for the contact picture.
Only the parameter name changed, from event to item.
